buletproofing delete method

+++ src/collaborative_platform/apps/files_management/api.py
@@ -188,10 +188,12 @@ def delete(request, **kwargs):
         dir = Directory.objects.filter(id=kwargs['directory_id']).get()
         log_activity(project=dir.project, user=request.user, related_dir=dir, action_text="deleted")
         dir.delete()
-    else:
+    elif 'file_id' in kwargs:
         file = File.objects.filter(id=kwargs['file_id']).get()
         log_activity(project=file.project, user=request.user, file=file, action_text="deleted")
         file.delete()
+    else:
+        return HttpResponseBadRequest("Invalid arguments")
     return HttpResponse("OK")
